GridPower now uses Doubles instead of Ints

+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/block/entity/CreativeMillBlockEntity.kt
@@ -11,6 +11,6 @@ class CreativeMillBlockEntity(
 	blockState: BlockState
 ) : GpSourceBlockEntity(ModBlockEntityTypes.CREATIVE_MILL.get(), pos, blockState) {
 
-	override fun getGpGeneration(): Int = ServerConfig.CONFIG.creativeMillGeneration.get()
+	override fun getGpGeneration(): Double = ServerConfig.CONFIG.creativeMillGeneration.get()
 
 }
